+++ CHANGELOG.md
@@ -37,6 +37,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#63](https://github.com/babylonlabs-io/covenant-emulator/pull/63) Add babylon
+address to keys command output and fix casing in dockerfile
+
 ## v0.10.0
 
 ### Improvements